+++ M3D2_Daniele_Dangela/backend-course_day5/src/main/java/com/mcs/be/course/service/CustomerService.java
@@ -1,11 +1,11 @@
 package com.mcs.be.course.service;
 
-import com.mcs.be.course.exception.ElementNotFound;
-import com.mcs.be.course.model.Article;
+
+import com.mcs.be.course.model.Customer;
 
 import java.util.List;
 
-public interface UserService 
+public interface CustomerService 
 {
    
 	  List<Customer> retrieveAllUsers();
